requirements.txt
psutil==5.9.8
openai==0.27.0
python-dotenv==1.0.0
tk